Updated how Date is displayed depending on whether we are in history screen or PendingOrders
